+++ src/main/java/cz/mendelu/xlinek/eduapp/api/subject/Subject.java
@@ -15,7 +15,7 @@ public class Subject {
     @GeneratedValue(strategy = GenerationType.TABLE, generator = "edu_sequence")
     @TableGenerator(name = "edu_sequence", table = "seq_table", pkColumnName = "entity", valueColumnName = "seq_value", initialValue = 0, allocationSize = 1)
     @Column(name = "idSubject")
-    private int idSubject;
+    private long idSubject;
     @Column(name = "title", nullable = false)
     private String title;
     @Column(name = "grade", nullable = false)
